fix: remove unnecessary echo statements and ensure proper file checks in scripts

+++ .bin/mv/jar.sh
@@ -1,3 +1,5 @@
-mv ./Floodgate-Spigot.jar plugins/
-mv ./Geyser-Spigot.jar plugins/
-mv ./Via-Version.jar plugins/
+if [ -f "./Floodgate-Spigot.jar" ] || [ -f "./Geyser-Spigot.jar" ] || [ -f "./Via-Version.jar" ]; then
+    mv ./Floodgate-Spigot.jar plugins/
+    mv ./Geyser-Spigot.jar plugins/
+    mv ./Via-Version.jar plugins/
+fi
